Added some comments to config. Fixed tickables storing. Moved AS and CI, started work on temperature.
