+++ src/main/java/com/github/evillootlye/caves/DangerousCaves.java
@@ -1,10 +1,11 @@
 package com.github.evillootlye.caves;
 
+import com.github.evillootlye.caves.caverns.AmbientSounds;
+import com.github.evillootlye.caves.caverns.CaveInsPlayerListener;
 import com.github.evillootlye.caves.configuration.Configuration;
 import com.github.evillootlye.caves.generator.CaveGenerator;
 import com.github.evillootlye.caves.mobs.DefaultMobs;
 import com.github.evillootlye.caves.mobs.MobsManager;
-import com.github.evillootlye.caves.surroundings.AmbientSounds;
 import org.bukkit.Bukkit;
 import org.bukkit.plugin.java.JavaPlugin;
