+++ src/main/java/com/github/evillootlye/caves/Dynamics.java
@@ -4,24 +4,28 @@ import org.bukkit.Bukkit;
 import org.bukkit.plugin.Plugin;
 
 import java.util.EnumMap;
+import java.util.HashSet;
 import java.util.Map;
+import java.util.Set;
 
 public class Dynamics {
-    private final Map<TickLevel, Tickable> tickables;
+    private final Map<TickLevel, Set<Tickable>> tickables;
 
     public Dynamics(Plugin plugin) {
         tickables = new EnumMap<>(TickLevel.class);
-        for(TickLevel level : TickLevel.values())
+        for(TickLevel level : TickLevel.values()) {
+            tickables.put(level, new HashSet<>());
             Bukkit.getScheduler().scheduleSyncRepeatingTask(
                     plugin,
-                    () -> tickables.get(level).tick(),
+                    () -> tickables.get(level).forEach(Tickable::tick),
                     level.ticks,
                     level.ticks
             );
+        }
     }
 
     public void subscribe(Tickable tick) {
-        tickables.put(tick.getTickLevel(), tick);
+        tickables.get(tick.getTickLevel()).add(tick);
     }
 
     public enum TickLevel {
